LayoutTests/inspector/styles-disable-inherited-expected.txt

Tests that disabling inherited style property does not break further style inspection.

Before disable
[collapsed] Style Attribute (element’s “style” attribute)

[expanded] div (user agent stylesheet)
display: block;

======== Inherited from div#container ========
[collapsed] Style Attribute (element’s “style” attribute)
font-weight: bold;


After disable:
[expanded] Style Attribute (element’s “style” attribute)

[expanded] div (user agent stylesheet)
display: block;
